+++ ch08/user-management-system/src/components/NavToggle.vue
@@ -1,0 +1,3 @@
+<template>
+  <v-app-bar-nav-icon @click="$emit('toggle-menu')" />
+</template>
